requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.1
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
requests>=2.31.0